Link to CTIS Methodology Report
The Methodology Report is meant to be the canonical source for information about the design of the survey, so we should provide a link to it here. Once the data is archived, we can change this to a link to the central documentation for the survey, including the methodology report, codebooks, and archival data.